+++ errors.go
@@ -13,4 +13,6 @@ var (
 	ErrProvider         = errors.New("provider error")
 	ErrProviderFetch    = errors.New("error fetching package")
 	ErrMigrateNeeded    = fmt.Errorf("migration needed. Call `%s migrate` to migrate files", os.Args[0])
+	ErrUnknownStateFileVersion = errors.New("unknown state file version")
+	ErrUnknownPackageFileVersion = errors.New("unknown state file version")
 )