+++ packages.go
@@ -3,14 +3,13 @@ package bpm
 import (
 	"os"
 	"runtime"
-	"strings"
 
 	"github.com/creasty/defaults"
 	"github.com/rs/zerolog"
 )
 
 const (
-	PackageSchemaVersion = 1
+	PackageSchemaVersion = 2
 )
 
 type PackageProvider interface {
@@ -19,6 +18,23 @@ type PackageProvider interface {
 }
 
 type Package struct {
+	PackageV2 `yaml:",inline"`
+}
+
+type PackageV2 struct {
+	SchemaVersion int    `yaml:"schema_version" default:"1"`
+	Name          string `yaml:"name"`
+	Provider      string `yaml:"provider"`
+	URL           string `yaml:"url"`
+	GOOS          map[string]string `yaml:"goos"`
+	GOARCH        map[string]string `yaml:"goarch"`
+	AssetPattern  string `yaml:"asset_pattern" default:"${goos}-${goarch}"`
+	ArchiveFormat string `yaml:"archive_format" default:""`
+	BinPattern    string `yaml:"bin_pattern" default:"${name}"`
+	DownloadUrl   string `yaml:"download_url" default:""`
+}
+
+type PackageV1 struct {
 	SchemaVersion int    `yaml:"schema_version" default:"1"`
 	Name          string `yaml:"name"`
 	Provider      string `yaml:"provider"`
@@ -32,8 +48,8 @@ type Package struct {
 }
 
 func (pkg *Package) SetDefaults() {
-	pkg.GOOS = strings.ToLower(runtime.GOOS)
-	pkg.GOARCH = strings.ToLower(runtime.GOARCH)
+	pkg.GOOS = make(map[string]string) //strings.ToLower(runtime.GOOS)
+	pkg.GOARCH = make(map[string]string) //strings.ToLower(runtime.GOARCH)
 }
 
 func (pkg *Package) UnmarshalYAML(unmarshal func(interface{}) error) error {
@@ -77,9 +93,17 @@ func (pkg *Package) patternExpand(pattern string, version string) string {
 	mapper := func(placeHolderName string) string {
 		switch placeHolderName {
 		case "goos":
-			return pkg.GOOS
+			goos, ok := pkg.GOOS[runtime.GOOS]
+			if ok {
+				return goos
+			}
+			return runtime.GOOS
 		case "goarch":
-			return pkg.GOARCH
+			goarch, ok := pkg.GOARCH[runtime.GOARCH]
+			if ok {
+				return goarch
+			}
+			return runtime.GOARCH
 		case "name":
 			return pkg.Name
 		case "version":