+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 * add outdated command
+* add migrate command
+* convert goos and goarch to maps to use same package definition on different platforms. This needs schema migration (package schema is now 2).
 
 ## 0.1.1 (2021-11-18)